# cmd/microvault/docs/swagger.yaml
basePath: /
definitions:
  controllers.ChallengeResponse:
    properties:
      challenge_id:
        type: string
      expires_at:
        type: integer
      transaction:
        type: string
    type: object
  controllers.VerifyRequest:
    properties:
      challenge_id:
        type: string
      signed_transaction:
        type: string
    required:
    - challenge_id
    - signed_transaction
    type: object
  controllers.VerifyResponse:
    properties:
      expires_at:
        type: integer
      token:
        type: string
    type: object
  fiber.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  github_com_Shamba-Records-Limited_microvault_pkg_payment_yellowcard.SettlementInfo:
    properties:
      cryptoAmount:
        type: number
      cryptoCurrency:
        type: string
      cryptoLocalRate:
        type: number
      cryptoNetwork:
        type: string
      cryptoUSDRate:
        type: number
      expiresAt:
        type: string
      lnInvoice:
        type: string
      walletAddress:
        type: string
      walletTag:
        type: string
    type: object
  github_com_Shamba-Records-Limited_microvault_pkg_payment_yellowcard.WebhookEvent:
    properties:
      data:
        $ref: '#/definitions/github_com_Shamba-Records-Limited_microvault_pkg_payment_yellowcard.WebhookPayload'
      event:
        type: string
      timestamp:
        type: string
    type: object
  github_com_Shamba-Records-Limited_microvault_pkg_payment_yellowcard.WebhookPayload:
    properties:
      amount:
        type: integer
      convertedAmount:
        type: integer
      country:
        type: string
      createdAt:
        type: string
      currency:
        type: string
      directSettlement:
        type: boolean
      id:
        type: string
      rate:
        type: number
      sequenceId:
        type: string
      settlementInfo:
        $ref: '#/definitions/github_com_Shamba-Records-Limited_microvault_pkg_payment_yellowcard.SettlementInfo'
      status:
        type: string
      updatedAt:
        type: string
    type: object
  middleware.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: smugane@shambarecords.com
    name: API Support
  description: A headless ERC-4626 tokenized vault engine for microlending built on
    the stellar network.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: http://swagger.io/terms/
  title: microvault API
  version: "1.0"
paths:
  /api/v1/auth/challenge:
    get:
      consumes:
      - application/json
      description: Generate a Stellar transaction challenge for authentication
      produces:
      - application/json
      responses:
        "200":
          description: Challenge generated successfully
          schema:
            $ref: '#/definitions/controllers.ChallengeResponse'
        "500":
          description: Failed to generate challenge
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: Get Authentication Challenge
      tags:
      - Authentication
  /api/v1/auth/verify:
    post:
      consumes:
      - application/json
      description: Verify a signed Stellar transaction challenge and receive a JWT
        token
      parameters:
      - description: Signed challenge verification
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.VerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Challenge verified successfully
          schema:
            $ref: '#/definitions/controllers.VerifyResponse'
        "400":
          description: Invalid request body or transaction
          schema:
            $ref: '#/definitions/middleware.Response'
        "401":
          description: Challenge verification failed
          schema:
            $ref: '#/definitions/middleware.Response'
        "404":
          description: Challenge not found
          schema:
            $ref: '#/definitions/middleware.Response'
        "500":
          description: Failed to generate token
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: Verify Challenge
      tags:
      - Authentication
  /api/v1/mobile/ussd/{provider}:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Handle USSD callback requests from the USSD gateway. The provider
        is specified in the URL path.
      parameters:
      - description: USSD provider name (e.g. africastalking)
        in: path
        name: provider
        required: true
        type: string
      - description: Session ID
        in: formData
        name: sessionId
        required: true
        type: string
      - description: User's phone number
        in: formData
        name: phoneNumber
        required: true
        type: string
      - description: User input text
        in: formData
        name: text
        type: string
      - description: USSD service code
        in: formData
        name: serviceCode
        required: true
        type: string
      - description: Network code
        in: formData
        name: networkCode
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: USSD response (CON/END)
          schema:
            type: string
        "500":
          description: END An error occurred. Please try again.
          schema:
            type: string
      summary: USSD Callback Handler
      tags:
      - USSD
  /api/v1/webhooks/yellowcard:
    post:
      consumes:
      - application/json
      description: Handle incoming webhooks from YellowCard payment service
      parameters:
      - description: HMAC signature for verification
        in: header
        name: X-YC-Signature
        required: true
        type: string
      - description: Webhook event data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Shamba-Records-Limited_microvault_pkg_payment_yellowcard.WebhookEvent'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Invalid webhook payload
          schema:
            $ref: '#/definitions/fiber.Error'
        "401":
          description: Signature verification failed
          schema:
            $ref: '#/definitions/fiber.Error'
        "500":
          description: Failed to process webhook
          schema:
            $ref: '#/definitions/fiber.Error'
      summary: Process YellowCard Payment Webhook
      tags:
      - Webhooks
swagger: "2.0"